Optimize Insertion Tail Latency in MyHash by Parameter Tuning
In this commit, several crucial modifications and optimizations have been made to enhance the performance of MyHash, which inherits from SepHash.
Firstly, the modified hash code has been migrated to MyHash, aiming to leverage and extend the existing functionality. To address the issue of high tail latency during insertion operations, specific adjustments were made to the ibv_modify_qp settings. By setting flags |= IBV_QP_MIN_RNR_TIMER and attr.min_rnr_timer = 3, the minimum time interval that the sender waits for the receiver to get ready after receiving an RNR - NAK (Receiver - Not - Ready Negative Acknowledgment) message is reduced. This is essential because when segments are full, they will be merged, and the server has to wait for the completion notification from the client after the merge to post recv, preventing overwriting of the existing data. Previously, when the client sent data and detected that the remote end hadn't received it, the network card would retry at intervals defined by min_rnr_timer. The overly long retry interval was a major contributor to the increased tail latency of insertion operations.
After these adjustments, although the tail latency is still higher than that of SepHash, the IOPS performance is comparable. Looking ahead, there are two potential optimization directions planned. One option is to abandon the built-in retry mechanism of the network card. Instead, if the client fails to receive an acknowledgment after sending data, it will immediately mark the operation as failed and then the client will take the initiative to retry, thus eliminating the delay caused by the network card's retries. The other idea is to divide CurSeg into multiple parts. Once a part is written, it will be merged and the recv for this part will be posted. Ideally, this enables continuous writing without encountering the "server not ready" situation. However, this approach may lead to incomplete clearing of the filter after merging, resulting in more false positives. To mitigate this, we plan to enforce a full merge and clear the filter at regular intervals in the future.